pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Анализатор страниц"
authors = ["Владимир Силин <sva@itx24.ru>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
flask = "^3.0.3"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.34.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
